--- a/packages/gatsby/src/query/query-runner.ts
+++ b/packages/gatsby/src/query/query-runner.ts
@@ -14,6 +14,7 @@ import errorParser from "./error-parser"
 import { GraphQLRunner } from "./graphql-runner"
 import { IExecutionResult, PageContext } from "./types"
 import { pageDataExists } from "../utils/page-data"
+import ldmbStore from "../utils/lmdb"
 
 const resultHashes = new Map()
 
@@ -181,7 +182,8 @@ export async function queryRunner(
         `json`,
         `${queryJob.id.replace(/\//g, `_`)}.json`
       )
-      await fs.outputFile(resultPath, resultJSON)
+      // await fs.outputFile(resultPath, resultJSON)
+      await ldmbStore.put(resultPath, resultJSON)
       store.dispatch({
         type: `ADD_PENDING_PAGE_DATA_WRITE`,
         payload: {