Testing/demoing an idea. This won't ever be merged.

For create-pages benchmark, this speeds up considerably the two fs IO heavy operations:

- query-running drops from from 11.5s to 4.2s
- writing page-data.json files drops from 17s to 5s.

Total time (to when just after writing page-data.json files to disk) drops from 43s to 23s.

One caveat to the above. This is measuring a drop in overhead. So the speed up for query running — 7 seconds / 100k queries — which is a big drop for create-pages which doesn nothing other than file IO essentially but would be less significant for comparatively more expensive query running e.g. markdown processing. So the relative drop will vary greatly. But every site should see around ~15-20s drop / 100k pages which is nothing to sneeze at.